Move Adapter\Decorator\ControlOverwrite to PreventOverwrite.
It seems more natural behavior to allow writes by default and then
having something to prevent them if needed. Contrary to preventing
writes on existing values that was the default behavior.

This simplified a lot more the code and provided a better API IMHO.
Not even to mention the tests, which became a lot more clearer and
objective.