Move Adapter\Decorator to Adapter\Rule.
Since they are not proper decorators, as they don't share always the
same interface. Expecting this behaviour just because of the name was
already annoying to me. And "Rule" proper reflect what future classes
should be in that namespace than "Decorators", that makes (at least for
me) very tempting to add an MultipleAdapters decorator in there. As for
a rule, it is clearly non-sense to add it there. #happypanda

Also moved the *.feature files of adapters to their own folder, mainly
because I don't like seeing things without a context.

Little changes to the tests were made to conform the "Rule" namespace
change. I really think that even the tests got better describe with
this.